SpringBootTask/src/test/java/com/example/springboottask/service/UserServiceImplTest.java
package com.example.springboottask.service;

import com.example.springboottask.converter.UserEntityConverter;
import com.example.springboottask.entity.UserEntity;
import com.example.springboottask.model.User;
import com.example.springboottask.repository.UserRepository;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import java.util.ArrayList;
import java.util.List;
import java.util.Optional;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.doReturn;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
class UserServiceImplTest {
    @Mock
    private UserRepository userRepository;

    @Mock
    UserEntityConverter userEntityConverter;

    @InjectMocks
    private UserServiceImpl userService;

    @Test
    void shouldReturnUserList() {
        List<UserEntity> expectedUserEntities = new ArrayList<>();
        expectedUserEntities.add(createUserEntity());
        List<User> expectedUsers = new ArrayList<>();
        expectedUsers.add(createUserModel());

        doReturn(expectedUserEntities).when(userRepository).findAll();
        when(userEntityConverter.toModel(expectedUserEntities)).thenReturn(expectedUsers);

        List<User> actualUsers = userService.getUsers();
        assertEquals(expectedUsers, actualUsers);
    }

    @Test
    void shouldSaveUserWhenUserIsGiven() {
        UserEntity userEntity = createUserEntity();

        when(userRepository.save(any(UserEntity.class))).thenReturn(userEntity);

        UserEntity savedUser = userRepository.save(userEntity);
        assertThat(savedUser.getFirstName()).isNotNull();
        assertThat(savedUser.getLastName()).isNotNull();
        assertThat(savedUser.getEmail()).isNotNull();
        assertThat(savedUser.getCity()).isNotNull();
    }
    
    @Test
    void shouldReturnUserWhenIdIsGiven() {
        UserEntity userEntity = new UserEntity();
        userEntity.setId(12L);
        User user = new User();
        user.setId(12L);

        when(userRepository.findById(userEntity.getId()))
                .thenReturn(Optional.of(userEntity));

        when(userEntityConverter.toModel(any(UserEntity.class))).thenReturn(user);

        User found = userService.getUser(userEntity.getId());
        assertThat(found.getId())
                .isEqualTo(userEntity.getId());
        // + verify
    }

    @Test
    void shouldDeleteUserWhenIdIsGiven() {
        Long userId = 12L;

        // create the instance we will test and give it our mock
        userService.deleteUser(12L);

        // check that the expected methods were called
        verify(userRepository).deleteById(userId);
    }

    private User createUserModel() {
        User user = new User();
        user.setFirstName("Sara");
        user.setLastName("Collins");
        user.setEmail("sara@mail.com");
        user.setCity("London");

        return user;
    }

    private UserEntity createUserEntity() {
        UserEntity userEntity = new UserEntity();
        userEntity.setFirstName("Sara");
        userEntity.setLastName("Collins");
        userEntity.setEmail("sara@mail.com");
        userEntity.setCity("London");

        return userEntity;
    }
}